Add files via upload
The project has still the parts of 1. sending email, 2. forgot password, 3. discount considered for the final price AND 4. some jquery working for some neater appearance. 
As of now the other goals from the user stories of this project have been successfully attained ! Thanks to the JAVA being self-debuggable and Ms. Heidi's coaching !